Check if the nginx and php config exists. Support domains.

+++ install.sh
@@ -4,16 +4,25 @@ export DEBIAN_FRONTEND="noninteractive"
 
 echo "Installing packages..."
 apt-get update -qq
-apt-get install -qq nginx php-fpm php-rrd curl zip net-tools
+apt-get install -qq nginx php-fpm php-rrd curl zip net-tools > /dev/null
 
-echo "Creating nginx config..."
-cat <<EOF > /etc/nginx/sites-enabled/default 
+echo "Discovering server IP..."
+server_ip="$(ifconfig | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -Eo '([0-9]*\.){3}[0-9]*' | grep -v '127.0.0.1')"
+echo "It's ${server_ip}"
+
+if [ -e /etc/nginx/sites-available/air-quality-info ]; then
+    echo "nginx config already exists, skipping..."
+else
+    read -e -p 'Enter the site domain (or press Enter to just use IP ${server_ip}): ' domain
+
+    echo "Creating nginx config..."
+    cat <<EOF > /etc/nginx/sites-available/air-quality-info
 server {
     listen 80;
 
-	server_name default_server;
-	root /var/www/air-quality-info;
-	index index.php;
+    server_name ${domain:-default_server};
+    root /var/www/air-quality-info;
+    index index.php;
 
     location / {
             try_files \$uri \$uri/ /index.php?\$args;
@@ -27,6 +36,20 @@ server {
     }
 }
 EOF
+    if [ -z "${domain}" ]; then
+        rm /etc/nginx/sites-enabled/default
+    fi
+fi
+
+if [ ! -L /etc/nginx/sites-enabled/air-quality-info ]; then
+    echo "Linking the nginx config..."
+    ln -s /etc/nginx/sites-available/air-quality-info /etc/nginx/sites-enabled
+fi
+
+domain="$(grep server_name /etc/nginx/sites-enabled/air-quality-info | awk '{ print $2 }' | tr -d ';')"
+if [ "${domain}" == 'default_server' ]; then
+    domain="${server_ip}"
+fi
 
 echo "Downloading Air Quality Info..."
 mkdir -p /var/www/air-quality-info
@@ -36,16 +59,19 @@ echo "Unpacking the htdocs"
 unzip -q /tmp/air-quality-info.zip -d /tmp
 mv /tmp/air-quality-info-master/htdocs/* /var/www/air-quality-info
 
-echo "Configuring Air Quality Info..."
-read -e -p    'Enter username: ' username
-read -e -s -p 'Enter password: ' password
-echo
-read -e -p    'Enter sensor id: ' sensor_id
+if [ -e /var/www/air-quality-info/config.php ]; then
+    echo "Air Quality Info config already exists, skipping..."
+else
+    echo "Configuring Air Quality Info..."
+    read -e -p    'Enter username: ' username
+    read -e -s -p 'Enter password: ' password
+    echo
+    read -e -p    'Enter sensor id: ' sensor_id
 
-cat <<EOF > /var/www/air-quality-info/config.php
+    cat <<EOF > /var/www/air-quality-info/config.php
 <?php
 define('CONFIG', array(
-  'devices' => array(
+'devices' => array(
     array(
         'user'        => '${username}',
         'password'    => '${password}',
@@ -53,14 +79,15 @@ define('CONFIG', array(
         'name'        => 'main',          # this will be used in URLs
         'description' => 'Main location', # user-friendly location name, will be used in navbar
     ),
-  ),
+),
 # Whether to store the last received JSON dump.
-  'store_json_payload' => true,
+'store_json_payload' => true,
 # Google Analytics ID
-  'ga_id' => ''
+'ga_id' => ''
 ));
 ?>
 EOF
+fi
 
 chown -R www-data:www-data /var/www/air-quality-info
 
@@ -68,14 +95,12 @@ echo "Starting PHP and nginx..."
 /etc/init.d/php7.0-fpm start
 /etc/init.d/nginx start
 
-server_ip="$(ifconfig | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -Eo '([0-9]*\.){3}[0-9]*' | grep -v '127.0.0.1')"
-
 echo "Installation is ready"
 echo ""
-echo "Page is available at http://${server_ip}"
+echo "Page is available at http://${domain}"
 echo ""
 echo "Configuration:"
-echo "Server: ${server_ip}"
+echo "Server: ${domain}"
 echo "Path: /main/update"
 echo "Port: 80"
 echo "User: ${username}"
